lib.model - remove all plaidml implementations
