+++ lib/model/session.py
@@ -8,15 +8,11 @@ from typing import Callable, ContextManager, List, Optional, Union
 import numpy as np
 import tensorflow as tf
 
-from lib.utils import get_backend
+# Ignore linting errors from Tensorflow's thoroughly broken import system
+from tensorflow.python.keras.layers import Activation  # pylint:disable=no-name-in-module
+from tensorflow.python.keras.models import load_model as k_load_model, Model  # noqa:E501  # pylint:disable=no-name-in-module
 
-if get_backend() == "amd":
-    from keras.layers import Activation
-    from keras.models import load_model as k_load_model, Model
-else:
-    # Ignore linting errors from Tensorflow's thoroughly broken import system
-    from tensorflow.keras.layers import Activation  # noqa pylint:disable=no-name-in-module,import-error
-    from tensorflow.keras.models import load_model as k_load_model, Model  # noqa pylint:disable=no-name-in-module,import-error
+from lib.utils import get_backend
 
 logger = logging.getLogger(__name__)  # pylint:disable=invalid-name
 
@@ -28,8 +24,7 @@ class KSession():
     actions performed on a model are handled consistently and can be performed in parallel in
     separate threads.
 
-    This is an early implementation of this class, and should be expanded out over time
-    with relevant `AMD`, `CPU` and `NVIDIA` backend methods.
+    This is an early implementation of this class, and should be expanded out over time.
 
     Notes
     -----
@@ -100,52 +95,13 @@ class KSession():
         """
         assert self._model is not None
         with self._context:
-            if self._backend == "amd" and batch_size is not None:
-                return self._amd_predict_with_optimized_batchsizes(feed, batch_size)
             return self._model.predict(feed, verbose=0, batch_size=batch_size)
-
-    def _amd_predict_with_optimized_batchsizes(
-            self,
-            feed: Union[List[np.ndarray], np.ndarray],
-            batch_size: int) -> Union[List[np.ndarray], np.ndarray]:
-        """ Minimizes the amount of kernels to be compiled when using the ``amd`` backend with
-        varying batch sizes while trying to keep the batchsize as high as possible.
-
-        Parameters
-        ----------
-        feed: numpy.ndarray or list
-            The feed to be provided to the model as input. This should be a ``numpy.ndarray``
-            for single inputs or a ``list`` of ``numpy.ndarray`` objects for multiple inputs.
-        batch_size: int
-            The upper batchsize to use.
-        """
-        assert self._model is not None
-        if isinstance(feed, np.ndarray):
-            feed = [feed]
-        items = feed[0].shape[0]
-        done_items = 0
-        results = []
-        while done_items < items:
-            if batch_size < 4:  # Not much difference in BS < 4
-                batch_size = 1
-            batch_items = ((items - done_items) // batch_size) * batch_size
-            if batch_items:
-                pred_data = [x[done_items:done_items + batch_items] for x in feed]
-                pred = self._model.predict(pred_data, batch_size=batch_size)
-                done_items += batch_items
-                results.append(pred)
-            batch_size //= 2
-        if isinstance(results[0], np.ndarray):
-            return np.concatenate(results)
-        return [np.concatenate(x) for x in zip(*results)]
 
     def _set_session(self,
                      allow_growth: bool,
                      exclude_gpus: list,
                      cpu_mode: bool) -> ContextManager:
         """ Sets the backend session options.
-
-        For AMD backend this does nothing.
 
         For CPU backends, this hides any GPUs from Tensorflow.
 
@@ -165,8 +121,6 @@ class KSession():
             ``True`` run the model on CPU. Default: ``False``
         """
         retval = nullcontext()
-        if self._backend == "amd":
-            return retval
         if self._backend == "cpu":
             logger.verbose("Hiding GPUs from Tensorflow")  # type:ignore
             tf.config.set_visible_devices([], "GPU")
@@ -201,8 +155,7 @@ class KSession():
         logger.verbose("Initializing plugin model: %s", self._name)  # type:ignore
         with self._context:
             self._model = k_load_model(self._model_path, compile=False, **self._model_kwargs)
-            if self._backend != "amd":
-                self._model.make_predict_function()
+            self._model.make_predict_function()
 
     def define_model(self, function: Callable) -> None:
         """ Defines a model from the given function.
@@ -233,8 +186,7 @@ class KSession():
         assert self._model is not None
         with self._context:
             self._model.load_weights(self._model_path)
-            if self._backend != "amd":
-                self._model.make_predict_function()
+            self._model.make_predict_function()
 
     def append_softmax_activation(self, layer_index: int = -1) -> None:
         """ Append a softmax activation layer to a model
